requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Settings
pydantic-settings==2.6.1

# Background tasks
celery==5.4.0
flower==2.0.1
redis==5.2.0

# Document parsing
pymupdf==1.24.14
python-docx==1.1.2
python-multipart==0.0.17

# ML / Vector search
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4

# LLM
openai==1.57.2
tenacity==9.0.0

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
